fix(train): use float32 for error_records metrics to support NPU hccl
PR #1035 (commit b2c1168) introduced `error_records_sum` as an int32
tensor and `error_records_total` as a default-dtype tensor. Both are
passed through `dist.reduce(op=ReduceOp.SUM)` in distributed training.

On Ascend NPU, the hccl backend does not support `ReduceOp.SUM` on
int32 tensors, causing a `HcclReduce` RuntimeError at training launch
on multi-NPU setups (Issue #1115). This is a regression from
v0.7.0.dev160 (c667ee1), which predates the error_records metrics.

Fix: change both tensors to float32, which is universally supported by
hccl/CUDA reduce operations.

Fixes #1115

+++ src/speculators/train/trainer.py
@@ -533,10 +533,12 @@ class Trainer:
             torch.nn.utils.clip_grad_norm_(self.model.parameters(), 1.0)
 
             metrics["error_records_sum"] = torch.tensor(
-                batch["error_records"], dtype=torch.int32, device=loss.device
+                batch["error_records"], dtype=torch.float32, device=loss.device
             )
             metrics["error_records_total"] = torch.tensor(
-                1.0 if self.rank == 0 else 0, device=loss.device
+                1.0 if self.rank == 0 else 0,
+                dtype=torch.float32,
+                device=loss.device,
             )
 
             timer.mark("bwd")
